requirements.txt
pandas==2.0.3
apache-airflow==2.7.0
apache-airflow-providers-apache-spark==4.1.0
pyspark==3.4.0
psycopg2-binary==2.9.6
boto3==1.26.0
pyarrow==12.0.0
kafka-python==2.0.2